% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFunctions.R
\name{atomCounts}
\alias{atomCounts}
\title{Count element occurrence}
\usage{
atomCounts(form)
}
\arguments{
\item{form}{Input formula}
}
\value{
data frame consisting of molecular symbol, element and count
}
\description{
Return a table of atomcounts from a chemical formula
}
