% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFunctions.R
\name{cas_check}
\alias{cas_check}
\title{Check cas number is correct}
\usage{
cas_check(cas)
}
\arguments{
\item{cas}{Input cas}
}
\value{
boolean
}
\description{
Check cas number is correct and return true if valid or false if not
}
