% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFunctions.R
\name{cas_to_smiles}
\alias{cas_to_smiles}
\title{return smiles string from cas}
\usage{
cas_to_smiles(cas)
}
\arguments{
\item{cas}{Input cas}
}
\value{
string
}
\description{
return a smiles string from a cas identifier using the NCI/CADD cactus service.
    If no smiles string is availble then NA is returned.
}
