% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFunctions.R
\name{data_from_pubchem}
\alias{data_from_pubchem}
\title{Extact data from PubChem}
\usage{
data_from_pubchem(cas)
}
\arguments{
\item{cas}{Input cas}
}
\value{
named list of properties
}
\description{
Extact data from PubChem using the cas number as a lookup.
    If more than one CID are matched, data for the highest ranked one are returned
}
