% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFunctions.R
\name{exactMass}
\alias{exactMass}
\title{Calculate exact mass}
\usage{
exactMass(form)
}
\arguments{
\item{form}{Input formula}
}
\value{
exact mass (numeric)
}
\description{
Calculate exact mass from formula
}
