% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemFunctions.R
\name{formulaSplit}
\alias{formulaSplit}
\title{Split chemical formula}
\usage{
formulaSplit(form)
}
\arguments{
\item{form}{Input formula}
}
\description{
Split a formula into chunks and return atom counts
}
